moved storage and deposits buffer to the common lib
