+++ common/credential-proxy/src/error.rs
@@ -1,4 +1,4 @@
-// Copyright 2024 Nym Technologies SA <contact@nymtech.net>
+// Copyright 2025 Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
 use nym_ecash_signer_check::SignerCheckError;
