+++ common/credential-proxy/src/lib.rs
@@ -1,0 +1,8 @@
+// Copyright 2025 Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub mod deposits_buffer;
+pub mod error;
+pub mod helpers;
+pub mod shared_state;
+pub mod storage;
