+++ common/credential-proxy/src/shared_state/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub mod nyxd_client;
+pub mod required_deposit_cache;
