+++ common/credential-proxy/src/storage/manager.rs
