+++ common/credential-proxy/src/storage/mod.rs
@@ -1,7 +1,6 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::credentials::ticketbook::NodeId;
 use crate::deposits_buffer::helpers::{BufferedDeposit, PerformedDeposits};
 use crate::error::CredentialProxyError;
 use crate::storage::manager::SqliteStorageManager;
@@ -26,6 +25,9 @@ use uuid::Uuid;
 
 mod manager;
 pub mod models;
+
+// TODO: proper import
+type NodeId = u64;
 
 #[derive(Clone)]
 pub struct CredentialProxyStorage {
@@ -74,7 +76,7 @@ impl CredentialProxyStorage {
     }
 
     #[allow(dead_code)]
-    pub(crate) async fn load_blinded_shares_status_by_shares_id(
+    pub async fn load_blinded_shares_status_by_shares_id(
         &self,
         id: i64,
     ) -> Result<Option<BlindedShares>, CredentialProxyError> {
@@ -84,7 +86,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn load_wallet_shares_by_shares_id(
+    pub async fn load_wallet_shares_by_shares_id(
         &self,
         id: i64,
     ) -> Result<Vec<MinimalWalletShare>, CredentialProxyError> {
@@ -94,7 +96,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn load_shares_error_by_shares_id(
+    pub async fn load_shares_error_by_shares_id(
         &self,
         id: i64,
     ) -> Result<Option<String>, CredentialProxyError> {
@@ -105,7 +107,7 @@ impl CredentialProxyStorage {
     }
 
     #[allow(dead_code)]
-    pub(crate) async fn load_blinded_shares_status_by_device_and_credential_id(
+    pub async fn load_blinded_shares_status_by_device_and_credential_id(
         &self,
         device_id: &str,
         credential_id: &str,
@@ -116,7 +118,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn load_wallet_shares_by_device_and_credential_id(
+    pub async fn load_wallet_shares_by_device_and_credential_id(
         &self,
         device_id: &str,
         credential_id: &str,
@@ -127,7 +129,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn load_shares_error_by_device_and_credential_id(
+    pub async fn load_shares_error_by_device_and_credential_id(
         &self,
         device_id: &str,
         credential_id: &str,
@@ -138,7 +140,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn insert_new_pending_async_shares_request(
+    pub async fn insert_new_pending_async_shares_request(
         &self,
         request: Uuid,
         device_id: &str,
@@ -150,7 +152,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn update_pending_async_blinded_shares_issued(
+    pub async fn update_pending_async_blinded_shares_issued(
         &self,
         available_shares: usize,
         device_id: &str,
@@ -166,7 +168,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn update_pending_async_blinded_shares_error(
+    pub async fn update_pending_async_blinded_shares_error(
         &self,
         available_shares: usize,
         device_id: &str,
@@ -184,7 +186,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn prune_old_blinded_shares(&self) -> Result<(), CredentialProxyError> {
+    pub async fn prune_old_blinded_shares(&self) -> Result<(), CredentialProxyError> {
         let max_age = OffsetDateTime::now_utc() - time::Duration::days(31);
 
         self.storage_manager
@@ -199,7 +201,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn insert_new_deposits(
+    pub async fn insert_new_deposits(
         &self,
         deposits: &PerformedDeposits,
     ) -> Result<(), CredentialProxyError> {
@@ -211,9 +213,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn load_unused_deposits(
-        &self,
-    ) -> Result<Vec<BufferedDeposit>, CredentialProxyError> {
+    pub async fn load_unused_deposits(&self) -> Result<Vec<BufferedDeposit>, CredentialProxyError> {
         self.storage_manager
             .load_unused_deposits()
             .await?
@@ -222,7 +222,7 @@ impl CredentialProxyStorage {
             .collect()
     }
 
-    pub(crate) async fn insert_deposit_usage(
+    pub async fn insert_deposit_usage(
         &self,
         deposit_id: DepositId,
         requested_on: OffsetDateTime,
@@ -240,7 +240,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn insert_deposit_usage_error(
+    pub async fn insert_deposit_usage_error(
         &self,
         deposit_id: DepositId,
         error: String,
@@ -251,7 +251,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn insert_partial_wallet_share(
+    pub async fn insert_partial_wallet_share(
         &self,
         deposit_id: DepositId,
         epoch_id: EpochId,
@@ -291,7 +291,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn get_master_verification_key(
+    pub async fn get_master_verification_key(
         &self,
         epoch_id: EpochId,
     ) -> Result<Option<EpochVerificationKey>, CredentialProxyError> {
@@ -309,7 +309,7 @@ impl CredentialProxyStorage {
         Ok(Some(deserialised))
     }
 
-    pub(crate) async fn insert_master_verification_key(
+    pub async fn insert_master_verification_key(
         &self,
         key: &EpochVerificationKey,
     ) -> Result<(), CredentialProxyError> {
@@ -320,7 +320,7 @@ impl CredentialProxyStorage {
             .await?)
     }
 
-    pub(crate) async fn get_master_coin_index_signatures(
+    pub async fn get_master_coin_index_signatures(
         &self,
         epoch_id: EpochId,
     ) -> Result<Option<AggregatedCoinIndicesSignatures>, CredentialProxyError> {
@@ -340,7 +340,7 @@ impl CredentialProxyStorage {
         Ok(Some(deserialised))
     }
 
-    pub(crate) async fn insert_master_coin_index_signatures(
+    pub async fn insert_master_coin_index_signatures(
         &self,
         signatures: &AggregatedCoinIndicesSignatures,
     ) -> Result<(), CredentialProxyError> {
@@ -355,7 +355,7 @@ impl CredentialProxyStorage {
         Ok(())
     }
 
-    pub(crate) async fn get_master_expiration_date_signatures(
+    pub async fn get_master_expiration_date_signatures(
         &self,
         expiration_date: Date,
         epoch_id: EpochId,
@@ -376,7 +376,7 @@ impl CredentialProxyStorage {
         Ok(Some(deserialised))
     }
 
-    pub(crate) async fn insert_master_expiration_date_signatures(
+    pub async fn insert_master_expiration_date_signatures(
         &self,
         signatures: &AggregatedExpirationDateSignatures,
     ) -> Result<(), CredentialProxyError> {
