+++ common/credential-proxy/src/storage/models.rs
