+++ nym-credential-proxy/nym-credential-proxy/src/tasks.rs
@@ -1,7 +1,7 @@
 // Copyright 2024 Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::storage::CredentialProxyStorage;
+use nym_credential_proxy_lib::storage::CredentialProxyStorage;
 use tokio_util::sync::CancellationToken;
 use tracing::{error, info};
